requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
autopep8==1.5.7
boto3==1.18.36
django-environ==0.5.0
django-filter==2.4.0
gunicorn==20.1.0
dj-database-url==0.5.0
django-heroku==0.3.1